introduced a ParallelEventHandlingConfigGroup with typed getters/setters; introduced a locking mechanism for the setters
git-svn-id: file:///srv/local/data/cleanRepo/trunk@20951 a0517364-e873-4167-b1c8-416daa5a6b91

+++ matsim/src/main/java/org/matsim/core/events/parallelEventsHandler/ParallelEventsManagerImpl.java
@@ -94,8 +94,8 @@ public class ParallelEventsManagerImpl extends EventsManagerImpl {
 	 *            Only use this constructor for larger simulations (20M+
 	 *            events).
 	 */
-	public ParallelEventsManagerImpl(int numberOfThreads, int estimatedNumberOfEvents) {
-		preInputBufferMaxLength = estimatedNumberOfEvents / 10;
+	public ParallelEventsManagerImpl(int numberOfThreads, long estimatedNumberOfEvents) {
+		preInputBufferMaxLength = (int) (estimatedNumberOfEvents / 10 );
 		init(numberOfThreads);
 	}
